fix: Ignore extra entries in `.env` file
When executed in another app's workspace, there can be a `.env` file for
the app. In that case, the content of the file is highly likely
incompatible with DLC's environment variable set. Since such usee case
is expected to be common, we should just ignore unknown keys inside the
file.